+++ realm/realm-library/src/main/cpp/io_realm_internal_Util.cpp
@@ -16,13 +16,13 @@
 
 #include <jni.h>
 
+#include "jni_util/jni_utils.hpp"
+
 #include <realm/string_data.hpp>
 #include <realm/unicode.hpp>
 
 #include "mem_usage.hpp"
 #include "util.hpp"
-
-#include "jni_util/jni_utils.hpp"
 
 using std::string;
 using namespace realm::jni_util;
@@ -53,6 +53,9 @@ JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void*)
         java_lang_double_init = env->GetMethodID(java_lang_double, "<init>", "(D)V");
         java_util_date = GetClass(env, "java/util/Date");
         java_util_date_init = env->GetMethodID(java_util_date, "<init>", "(J)V");
+#if REALM_ENABLE_SYNC
+        java_syncmanager = GetClass(env, "io/realm/SyncManager");
+#endif
     }
 
     return JNI_VERSION_1_6;
@@ -70,6 +73,9 @@ JNIEXPORT void JNI_OnUnload(JavaVM* vm, void*)
         env->DeleteGlobalRef(java_lang_double);
         env->DeleteGlobalRef(java_util_date);
         env->DeleteGlobalRef(java_lang_string);
+#if REALM_ENABLE_SYNC
+        env->DeleteGlobalRef(java_syncmanager);
+#endif
     }
 }