+++ realm/realm-library/src/objectServer/java/io/realm/internal/SyncObjectServerFacade.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.realm.internal.objectserver;
+package io.realm.internal;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
@@ -25,12 +25,9 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 
 import io.realm.RealmConfiguration;
-import io.realm.SyncSession;
 import io.realm.SyncConfiguration;
 import io.realm.SyncManager;
 import io.realm.exceptions.RealmException;
-import io.realm.internal.Keep;
-import io.realm.internal.ObjectServerFacade;
 import io.realm.internal.network.NetworkStateReceiver;
 
 @SuppressWarnings({"unused", "WeakerAccess"}) // Used through reflection. See ObjectServerFacade
@@ -41,6 +38,7 @@ public class SyncObjectServerFacade extends ObjectServerFacade {
             "'configuration' has to be an instance of 'SyncConfiguration'.";
     @SuppressLint("StaticFieldLeak") //
     private static Context applicationContext;
+    private static volatile Method removeSessionMethod;
 
     @Override
     public void init(Context context) {
@@ -71,33 +69,12 @@ public class SyncObjectServerFacade extends ObjectServerFacade {
     }
 
     @Override
-    public void notifyCommit(RealmConfiguration configuration, long lastSnapshotVersion) {
-        if (configuration instanceof SyncConfiguration) {
-            SyncSession publicSession = SyncManager.getSession((SyncConfiguration) configuration);
-            ObjectServerSession session = SessionStore.getPrivateSession(publicSession);
-            session.notifyCommit(lastSnapshotVersion);
-        } else {
-            throw new IllegalArgumentException(WRONG_TYPE_OF_CONFIGURATION);
-        }
-    }
-
-    @Override
     public void realmClosed(RealmConfiguration configuration) {
+        // Last Thread using the specified configuration is closed
+        // delete the wrapped Java session
         if (configuration instanceof SyncConfiguration) {
-            SyncSession publicSession = SyncManager.getSession((SyncConfiguration) configuration);
-            ObjectServerSession session = SessionStore.getPrivateSession(publicSession);
-            session.getSyncPolicy().onRealmClosed(session);
-        } else {
-            throw new IllegalArgumentException(WRONG_TYPE_OF_CONFIGURATION);
-        }
-    }
-
-    @Override
-    public void realmOpened(RealmConfiguration configuration) {
-        if (configuration instanceof SyncConfiguration) {
-            SyncSession publicSession = SyncManager.getSession((SyncConfiguration) configuration);
-            ObjectServerSession session = SessionStore.getPrivateSession(publicSession);
-            session.getSyncPolicy().onRealmOpened(session);
+            SyncConfiguration syncConfig = (SyncConfiguration) configuration;
+            invokeRemoveSession(syncConfig);
         } else {
             throw new IllegalArgumentException(WRONG_TYPE_OF_CONFIGURATION);
         }
@@ -108,14 +85,48 @@ public class SyncObjectServerFacade extends ObjectServerFacade {
         if (config instanceof SyncConfiguration) {
             SyncConfiguration syncConfig = (SyncConfiguration) config;
             String rosServerUrl = syncConfig.getServerUrl().toString();
-            String rosUserToken = syncConfig.getUser().getAccessToken();
-            return new String[]{rosServerUrl, rosUserToken};
+            String rosUserIdentity = syncConfig.getUser().getIdentity();
+            String syncRealmAuthUrl = syncConfig.getUser().getAuthenticationUrl().toString();
+            String rosRefreshToken = syncConfig.getUser().getAccessToken().value();
+            return new String[]{rosUserIdentity, rosServerUrl, syncRealmAuthUrl, rosRefreshToken};
         } else {
-            return new String[2];
+            return new String[4];
         }
     }
 
-    static Context getApplicationContext() {
+    public static Context getApplicationContext() {
         return applicationContext;
+    }
+
+    @Override
+    public void wrapObjectStoreSessionIfRequired(RealmConfiguration config) {
+        if (config instanceof SyncConfiguration) {
+            SyncManager.getSession((SyncConfiguration) config);
+        }
+    }
+
+    //FIXME remove this reflection call once we redesign the SyncManager to separate interface
+    //      from implementation to avoid issue like exposing internal method like SyncManager#removeSession
+    //      or SyncSession#close. This happens because SyncObjectServerFacade is internal, whereas
+    //      SyncManager#removeSession or SyncSession#close are package private & should not be public.
+    private void invokeRemoveSession(SyncConfiguration syncConfig) {
+        try {
+            if (removeSessionMethod == null) {
+                synchronized (SyncObjectServerFacade.class) {
+                    if (removeSessionMethod == null) {
+                        Method removeSession = SyncManager.class.getDeclaredMethod("removeSession", SyncConfiguration.class);
+                        removeSession.setAccessible(true);
+                        removeSessionMethod = removeSession;
+                    }
+                }
+            }
+            removeSessionMethod.invoke(null, syncConfig);
+        } catch (NoSuchMethodException e) {
+            throw new RealmException("Could not lookup method to remove session: " + syncConfig.toString(), e);
+        } catch (InvocationTargetException e) {
+            throw new RealmException("Could not invoke method to remove session: " + syncConfig.toString(), e);
+        } catch (IllegalAccessException e) {
+            throw new RealmException("Could not remove session: " + syncConfig.toString(), e);
+        }
     }
 }